[TM-2581] Get feedback and answers from the current update request if there is one.
The client does not need to care about if there's an active update request or not - with this, it can just always depend on feedback, feedbackFields and answers from this DTO instead of figuring out if there is an active update request or not itself. This also avoids doing a bunch of work in the DB to collect the current answers if there is a current update request.